Full-Screen Background in Proper Container, Fixed Package.json Caching
In detail the full-screen background has moved to the "under-main-view" container which actually does what it's meant to do (which is under the main view).

This solves a bunch of other issues without any other work like the full-screen background extending past the bounds of the actual visible rounded container.

Package.json also doesn't cache anymore upon request - this was an issue as JSDelivr could update its cache but locally we're still using an older cache!

Fixes #3, Fixes #4