Updated to check code instead of package.json
The reason for this is because we already bundle in the version property into the main code anyways.

It's possible that the cache on JSDelivr for both update at separate times as well - it's just better that we check the main cached thing that will actually determine if they get the latest update or not.