Implemented new Player service that centralizes Song Information/Playback
This is a significant part of the whole integration of the new lyrics. This brings in a readable playback state for the songs timestamp, general metadata/information, and parsed-lyrics.

Which is another change which is that we now store the parsed lyrics INSTEAD of the raw lyrics-result itself.

The LyricsParser.ts file was brought over directly from my local build. The rest of the stuff will follow suit with very minor adjustments needed to have it fit within the extension.