Made Lyric Fetching Faster
Most of the time Apple will have lyrics that are not static. In this case we should just completely skip getting the Spotify lyrics. So, what I've done is made it so that they don't run at the same time. Instead we run Backend lyrics first and then determine if we need to run Spotify lyric-fetching (if we are either missing lyrics from the backend or they are static synced). We only return Spotify lyrics IF we can find synced ones - otherwise we just return no lyrics (because of course Spotify has incorrect lyrics for songs when they are statically synced)