Fixed Alignment Detection
Main issue was that we were parsing line-by-line and in Apple Musics case we were interpreting the body HTML - which is not good because that's a lot of characters that aren't related to the lyrics at all. So we construct a text-based view of the lyrics and then process that. Therefore the language will always be based off what the song mostly is!