Removed Malay as Right-to-Left language
Malay can be written in either Latin (Rumi) or Arabic (Jawi) script. When I initially searched the right-to-left languages "zlm" (Malay) was listed as one of them only because the Jawi script was in Arabic (a right-to-left script/language).

Obviously mostly everything now uses Rumi and is easily identifiable as that.

Also the Jawi version will be recognized as Arabic anyways so we're not losing anything.